+++ docs/source/generate_rst.py
@@ -5,7 +5,7 @@ from os.path import join as pj
 
 
 class RSTGenerator:
-    title_line = "=" * 50
+    title_line = "=" * 80
     toctree = ".. toctree::\n    {}"
     automodule = ".. automodule:: {}\n    :members:\n    :undoc-members:\n    :show-inheritance:\n    :inherited-members:"
 
@@ -21,11 +21,14 @@ class RSTGenerator:
     def cleanup(self):
         if os.path.exists("modules.rst"):
             os.remove("modules.rst")
-        shutil.rmtree(self.dfolder)
+        try:
+            shutil.rmtree(self.dfolder)
+        except FileNotFoundError:
+            pass
         os.makedirs(self.dfolder)
 
     def write(self, path, content):
-        if type(content) == type([]):
+        if isinstance(content, list):
             content = "\n".join(content)
 
         with open(path, "w") as f:
@@ -33,70 +36,68 @@ class RSTGenerator:
 
         print(f"Finish writing {path}")
 
-    def single_file_module(self):
-        """Process the module in the self.pfolder/*.py"""
-
-        for module_name in glob.glob(pj(self.pfolder, "*.py")):
+    def _file_generate(self, package_parents):
+        file_list = []
+        for module_name in glob.glob(pj(self.pfolder, *package_parents, "*.py")):
             module_name = os.path.basename(module_name)[:-3]
             if module_name in self.ingnored_modules:
                 continue
 
-            rst_file = pj(self.dfolder, f"{module_name}.rst")
+            rst_file = pj(self.dfolder, *package_parents, f"{module_name}.rst")
+            name = f"{self.name}"
+            for n in package_parents:
+                name += f".{n}"
+            name += f".{module_name}"
             content = [
-                f"{self.name}.{module_name}",
+                name,
                 self.title_line,
-                self.automodule.format(f"{self.name}.{module_name}"),
+                self.automodule.format(name),
             ]
-
             self.write(rst_file, content)
-            self.tree[rst_file] = []
-
-    def subdir_files_module(self):
-        """Write the rst files for modules with subdir or files"""
-        for subdir in glob.glob(pj(self.pfolder, "*/")):
+            if not package_parents:
+                upper = self.dfolder
+            else:
+                upper = package_parents[-1]
+            file_list.append(upper + f"/{module_name}.rst")
+        for subdir in glob.glob(pj(self.pfolder, *package_parents, "*/")):
             if "_" in subdir:
                 continue
-
             subdir = os.path.basename(os.path.normpath(subdir))
-            os.makedirs(pj(self.dfolder, subdir), exist_ok=True)
-            rst_file = pj(self.dfolder, f"{subdir}.rst")
-            self.tree[rst_file] = []
+            os.makedirs(pj(self.dfolder, *package_parents, subdir), exist_ok=True)
+            rst_file = pj(self.dfolder, *package_parents, f"{subdir}.rst")
+            subdir_filelist = self._file_generate(package_parents + [subdir])
 
-            for module_name in glob.glob(pj(self.pfolder, subdir, f"*.py")):
-                module_name = os.path.basename(module_name)[:-3]
-                if module_name in self.ingnored_modules:
-                    continue
-
-                content = [
-                    f"{self.name}.{subdir}.{module_name}",
-                    self.title_line,
-                    self.automodule.format(f"{self.name}.{subdir}.{module_name}"),
-                ]
-
-                self.write(pj(self.dfolder, subdir, f"{module_name}.rst"), content)
-                self.tree[rst_file].append(f"{subdir}/{module_name}.rst")
-
+            name = f"{self.name}"
+            for n in package_parents:
+                name += f".{n}"
+            name += f".{subdir}.rst"
             content = [
-                f"{self.name}.{subdir}",
+                name,
                 self.title_line,
-                self.toctree.format("\n    ".join(sorted(self.tree[rst_file]))),
+                self.toctree.format("\n    ".join(sorted(subdir_filelist))),
             ]
             self.write(rst_file, content)
 
-    def modules_file(self):
+            if not package_parents:
+                upper = self.dfolder
+            else:
+                upper = package_parents[-1]
+            file_list.append(upper + f"/{subdir}.rst")
+        return file_list
+
+    def modules_file(self, file_list):
         """Write the modules.rst"""
         content = [
             self.name,
             self.title_line,
-            self.toctree.format("\n    ".join(sorted(self.tree.keys()))),
+            self.toctree.format("\n    ".join(sorted(file_list))),
         ]
         self.write("modules.rst", content)
 
     def start(self):
         self.cleanup()
-        self.single_file_module()
-        self.subdir_files_module()
-        self.modules_file()
+        file_list = self._file_generate([])
+        self.modules_file(file_list)
 
 
 if __name__ == "__main__":